# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for quiz game with host management and Telegram bot integration",
        "title": "Quiz Game API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticate host and return JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login as host",
                "parameters": [
                    {
                        "description": "Login data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.AuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Create a new host account and return JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new host",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sessions/join": {
            "post": {
                "description": "Join a session by code with telegram ID and nickname",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participants"
                ],
                "summary": "Join a quiz session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bot API Key",
                        "name": "X-Bot-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Join data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.JoinSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz-game-backend_internal_services.JoinResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sessions/{id}/answer": {
            "post": {
                "description": "Submit an answer to the current question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participants"
                ],
                "summary": "Submit an answer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bot API Key",
                        "name": "X-Bot-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Answer data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.SubmitAnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sessions/{id}/my-result": {
            "get": {
                "description": "Get whether the participant answered correctly and their score",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participants"
                ],
                "summary": "Get participant result for current question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bot API Key",
                        "name": "X-Bot-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Telegram ID",
                        "name": "telegram_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz-game-backend_internal_services.ParticipantResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handlers.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
        "internal_handlers.JoinSessionRequest": {
            "type": "object",
            "required": [
                "code",
                "nickname",
                "telegram_id"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "123456"
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Player1"
                },
                "telegram_id": {
                    "type": "integer",
                    "example": 123456789
                }
            }
        },
        "internal_handlers.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "host1"
                }
            }
        },
        "internal_handlers.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        },
...
        "internal_handlers.RegisterRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "host1"
                }
            }
        },
        "internal_handlers.SubmitAnswerRequest": {
            "type": "object",
            "required": [
                "option_id",
                "telegram_id"
            ],
            "properties": {
                "option_id": {
                    "type": "integer",
                    "example": 1
                },
                "telegram_id": {
                    "type": "integer",
                    "example": 123456789
                }
            }
        },
...
        "quiz-game-backend_internal_models.Participant": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "joined_at": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "session_id": {
                    "type": "integer"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "total_score": {
                    "type": "integer"
                }
            }
        },
...
        "quiz-game-backend_internal_services.JoinResult": {
            "type": "object",
            "properties": {
                "participant": {
                    "$ref": "#/definitions/quiz-game-backend_internal_models.Participant"
                },
                "session_id": {
                    "type": "integer"
                }
            }
        },
...
        "quiz-game-backend_internal_services.ParticipantResult": {
            "type": "object",
            "properties": {
                "answered": {
                    "type": "boolean"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "option_id": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "total_score": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}